+++ contrib/platform/test/com/sun/jna/platform/mac/SystemBTest.java
@@ -16,19 +16,32 @@
  */
 package com.sun.jna.platform.mac;
 
-import junit.framework.TestCase;
-
-import com.sun.jna.platform.mac.SystemB.HostCpuLoadInfo;
-import com.sun.jna.platform.mac.SystemB.HostLoadInfo;
-import com.sun.jna.platform.mac.SystemB.VMStatistics;
-import com.sun.jna.platform.mac.SystemB.VMStatistics64;
-
 import com.sun.jna.Memory;
+import com.sun.jna.Native;
 import com.sun.jna.Platform;
 import com.sun.jna.Pointer;
+import com.sun.jna.platform.mac.SystemB.Group;
+import com.sun.jna.platform.mac.SystemB.HostCpuLoadInfo;
+import com.sun.jna.platform.mac.SystemB.HostLoadInfo;
+import com.sun.jna.platform.mac.SystemB.IFmsgHdr;
+import com.sun.jna.platform.mac.SystemB.IFmsgHdr2;
+import com.sun.jna.platform.mac.SystemB.Passwd;
+import com.sun.jna.platform.mac.SystemB.ProcTaskAllInfo;
+import com.sun.jna.platform.mac.SystemB.RUsageInfoV2;
+import com.sun.jna.platform.mac.SystemB.Statfs;
+import com.sun.jna.platform.mac.SystemB.Timeval;
+import com.sun.jna.platform.mac.SystemB.Timezone;
+import com.sun.jna.platform.mac.SystemB.VMMeter;
+import com.sun.jna.platform.mac.SystemB.VMStatistics;
+import com.sun.jna.platform.mac.SystemB.VMStatistics64;
+import com.sun.jna.platform.mac.SystemB.VnodeInfoPath;
+import com.sun.jna.platform.mac.SystemB.VnodePathInfo;
+import com.sun.jna.platform.mac.SystemB.XswUsage;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.LongByReference;
 import com.sun.jna.ptr.PointerByReference;
+
+import junit.framework.TestCase;
 
 /**
  * Exercise the {@link SystemB} class.
@@ -169,8 +182,155 @@ public class SystemBTest extends TestCase {
       assertTrue(loadavg[2] >= 0);
     }
 	
-	public static void main(java.lang.String[] argList) {
-		junit.textui.TestRunner.run(SystemBTest.class);
-	}
+    public void testTimeofDay() {
+        Timeval tp = new Timeval();
+        long before = System.currentTimeMillis();
+        assertEquals(0, SystemB.INSTANCE.gettimeofday(tp, null));
+        long timeofday = tp.tv_sec.longValue() * 1000L + tp.tv_usec / 1000;
+        long after = System.currentTimeMillis();
+        assertTrue(before <= timeofday);
+        assertTrue(after >= timeofday);
 
+        Timezone tzp = new Timezone();
+        assertEquals(0, SystemB.INSTANCE.gettimeofday(null, tzp));
+        // DST is 0 or 1
+        assertTrue(tzp.tz_dsttime <= 1);
+        // Timezones are GMT -12 (west) to +14 hours
+        assertTrue(tzp.tz_minuteswest <= 12 * 60);
+        assertTrue(tzp.tz_minuteswest >= -14 * 60);
+    }
+
+    public void testVMMeter() {
+        int machPort = SystemB.INSTANCE.mach_host_self();
+        VMMeter vmstats = new VMMeter();
+        assertEquals(0, SystemB.INSTANCE.host_statistics(machPort, SystemB.HOST_VM_INFO, vmstats,
+                new IntByReference(vmstats.size())));
+        assertTrue(vmstats.v_lookups >= 0);
+    }
+
+    public void testStatfs() {
+        // Use statfs to populate mount point map
+        int numfs = SystemB.INSTANCE.getfsstat64(null, 0, 0);
+        Statfs[] fs = new Statfs[numfs];
+        // Fill array with results
+        SystemB.INSTANCE.getfsstat64(fs, numfs * new Statfs().size(), SystemB.MNT_NOWAIT);
+        // Iterate all mounted file systems
+        for (Statfs f : fs) {
+            assertTrue(f.f_bfree <= f.f_blocks);
+            assertTrue(f.f_ffree <= f.f_files);
+        }
+    }
+
+    public void testXswUsage() {
+        XswUsage xswUsage = new XswUsage();
+        assertEquals(0, SystemB.INSTANCE.sysctlbyname("vm.swapusage", xswUsage.getPointer(),
+                new IntByReference(xswUsage.size()), null, 0));
+        xswUsage.read();
+        assertTrue(xswUsage.xsu_used <= xswUsage.xsu_total);
+    }
+
+    public void testProcessStructures() {
+        // Calc max # of processes
+        IntByReference size = new IntByReference(4);
+        Memory mem = new Memory(4);
+        assertEquals(0, SystemB.INSTANCE.sysctlbyname("kern.maxproc", mem, size, null, 0));
+        int maxProc = mem.getInt(0);
+
+        // Get list of pids
+        int[] pids = new int[maxProc];
+        int bytesReturned = SystemB.INSTANCE.proc_listpids(SystemB.PROC_ALL_PIDS, 0, pids, maxProc * 4);
+        assertTrue(bytesReturned > 0);
+        assertEquals(0, bytesReturned % 4);
+
+        // Current pid should be in list
+        int pid = SystemB.INSTANCE.getpid();
+        assertTrue(pid > 0);
+        boolean foundPid = false;
+        for (int i = 0; i < pids.length; i++) {
+            if (pids[i] == pid) {
+                foundPid = true;
+                break;
+            }
+        }
+        assertTrue(foundPid);
+
+        // ProcTaskAllInfo
+        ProcTaskAllInfo taskAllInfo = new ProcTaskAllInfo();
+        bytesReturned = SystemB.INSTANCE.proc_pidinfo(pid, SystemB.PROC_PIDTASKALLINFO, 0, taskAllInfo,
+                taskAllInfo.size());
+        assertTrue(bytesReturned > 0);
+        assertEquals(0, bytesReturned % Native.getNativeSize(ProcTaskAllInfo.class));
+
+        Pointer buf = new Memory(SystemB.PROC_PIDPATHINFO_MAXSIZE);
+        bytesReturned = SystemB.INSTANCE.proc_pidpath(pid, buf, SystemB.PROC_PIDPATHINFO_MAXSIZE);
+        String path = buf.getString(0).trim();
+        assertEquals(bytesReturned, path.length());
+
+        // ProcTaskInfo
+        assertTrue(taskAllInfo.ptinfo.pti_threadnum > 0);
+
+        // ProcBsdInfo
+        Passwd user = SystemB.INSTANCE.getpwuid(taskAllInfo.pbsd.pbi_uid);
+        assertEquals(user.pw_uid, taskAllInfo.pbsd.pbi_uid);
+
+        Group group = SystemB.INSTANCE.getgrgid(taskAllInfo.pbsd.pbi_gid);
+        assertEquals(group.gr_gid, taskAllInfo.pbsd.pbi_gid);
+
+        RUsageInfoV2 rUsageInfoV2 = new RUsageInfoV2();
+        assertEquals(0, SystemB.INSTANCE.proc_pid_rusage(pid, SystemB.RUSAGE_INFO_V2, rUsageInfoV2));
+        assertTrue(rUsageInfoV2.ri_diskio_bytesread >= 0);
+
+        VnodePathInfo vpi = new VnodePathInfo();
+        bytesReturned = SystemB.INSTANCE.proc_pidinfo(pid, SystemB.PROC_PIDVNODEPATHINFO, 0, vpi, vpi.size());
+        assertTrue(bytesReturned > 0);
+        assertEquals(0, bytesReturned % Native.getNativeSize(VnodeInfoPath.class));
+
+        String cwd = Native.toString(vpi.pvi_cdir.vip_path);
+        assertTrue(cwd.length() > 0);
+    }
+
+    public void testIFs() {
+        int RTM_IFINFO2 = 0x12;
+        // Get buffer of all interface information
+        int CTL_NET = 4;
+        int PF_ROUTE = 17;
+        int NET_RT_IFLIST2 = 6;
+        int[] mib = { CTL_NET, PF_ROUTE, 0, 0, NET_RT_IFLIST2, 0 };
+
+        IntByReference len = new IntByReference();
+        assertEquals(0, SystemB.INSTANCE.sysctl(mib, 6, null, len, null, 0));
+        // Add enough room for max size of IFmsgHdr to avoid JNA bounds check
+        // problems with worst case structure size
+        Memory buf = new Memory(len.getValue() + 112);
+        assertEquals(0, SystemB.INSTANCE.sysctl(mib, 6, buf, len, null, 0));
+
+        // Iterate offset from buf's pointer up to limit of buf
+        int lim = len.getValue();
+        int next = 0;
+        while (next < lim) {
+            // Get pointer to current native part of buf
+            Pointer p = buf.share(next);
+            // Cast pointer to if_msghdr
+            IFmsgHdr ifm = new IFmsgHdr(p);
+            ifm.read();
+            // Advance next
+            next += ifm.ifm_msglen;
+            // Skip messages which are not the right format
+            if (ifm.ifm_type != RTM_IFINFO2) {
+                continue;
+            }
+            // Cast pointer to if_msghdr2
+            IFmsgHdr2 if2m = new IFmsgHdr2(p);
+            if2m.read();
+
+
+            assertTrue(if2m.ifm_index >= 0);
+            assertTrue(if2m.ifm_data.ifi_ibytes >= 0);
+            assertTrue(if2m.ifm_data.ifi_lastchange.tv_usec >= 0);
+        }
+    }
+
+    public static void main(java.lang.String[] argList) {
+        junit.textui.TestRunner.run(SystemBTest.class);
+    }
 }